+++ src/WyzeSmartHome.js
@@ -9,9 +9,6 @@ const WyzeMotionSensor = require('./accessories/WyzeMotionSensor')
 const WyzeTemperatureHumidity = require('./accessories/WyzeTemperatureHumidity')
 const WyzeLeakSensor = require('./accessories/WyzeLeakSensor')
 const WyzeCamera = require('./accessories/WyzeCamera')
-const WyzeSwitch = require('./accessories/WyzeSwitch')
-const WyzeThermostat = require('./accessories/WyzeThermostat')
-const WyzeHMS = require('./accessories/WyzeHMS')
 
 const PLUGIN_NAME = 'homebridge-wyze-smart-home'
 const PLATFORM_NAME = 'WyzeSmartHome'
@@ -87,7 +84,6 @@ module.exports = class WyzeSmartHome {
       const accessory = await this.loadDevice(device, timestamp)
       if (accessory) {
         foundAccessories.push(accessory)
-        this.api.updatePlatformAccessories([accessory])
       }
     }
 
@@ -102,10 +98,9 @@ module.exports = class WyzeSmartHome {
   }
 
   async loadDevice(device, timestamp) {
-
     const accessoryClass = this.getAccessoryClass(device.product_type, device.product_model)
     if (!accessoryClass) {
-     this.log.debug(`Unsupported device type or device is ignored: ${device.product_type} (Model: ${device.product_model})`)
+      this.log.debug(`Unsupported device type or device is ignored: ${device.product_type} (Model: ${device.product_model})`)
       return
     }
 
@@ -123,24 +118,12 @@ module.exports = class WyzeSmartHome {
       const homeKitAccessory = this.createHomeKitAccessory(device)
       accessory = new accessoryClass(this, homeKitAccessory)
       this.accessories.push(accessory)
+    } else {
+      this.log.debug(`Loading accessory from cache ${device.nickname} (MAC: ${device.mac})`)
     }
-    accessory.update(device, timestamp)    
+      accessory.update(device, timestamp)    
 
     return accessory
-  }
-
-  async loadHmsDevice() {
-    // Load HMSID
-    let hmsID
-    const response = await this.client.getPlanBindingListByUser()
-    hmsID = response.data[0].deviceList[0].device_id
-    console.log(hmsID)
-    let list = response.data[0].deviceList[0]
-    console.log(list)
-    let list1 = response.data[0]
-    console.log(list1)
-
-
   }
 
   getAccessoryClass(type, model) {
@@ -168,13 +151,6 @@ module.exports = class WyzeSmartHome {
       case 'Camera':
         if (model === 'WYZEDB3') return
         return WyzeCamera
-      case 'Common':
-        if (model === 'JA_HP') return
-        return WyzeSwitch
-      case 'Thermostat':
-        return WyzeThermostat
-      case 'S1Gateway':
-        return WyzeHMS
     }
   }
 
@@ -182,13 +158,13 @@ module.exports = class WyzeSmartHome {
     const uuid = UUIDGen.generate(device.mac)
 
     const homeKitAccessory = new Accessory(device.nickname, uuid)
-      homeKitAccessory.context = {
-        mac: device.mac,
-        product_type: device.product_type,
-        product_model: device.product_model,
-        nickname: device.nickname,
-        conn_state: device.conn_state
-      }
+
+    homeKitAccessory.context = {
+      mac: device.mac,
+      product_type: device.product_type,
+      product_model: device.product_model,
+      nickname: device.nickname
+    }
 
     this.api.registerPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [homeKitAccessory])
     return homeKitAccessory
@@ -208,7 +184,6 @@ module.exports = class WyzeSmartHome {
       this.accessories.push(accessory)
     } else {
       this.api.unregisterPlatformAccessories(PLUGIN_NAME, PLATFORM_NAME, [homeKitAccessory])
-      this.log.debug(`Loading accessory from cache ${device.nickname} (MAC: ${device.mac})`)
     }
   }
 }
