--- ~/Work/homebridge-wyze-api/accessories/cachedAccessories
@@ -1,1 +1,0 @@
-[{"plugin":"homebridge-wyze-smart-home","platform":"WyzeSmartHome","context":{"mac":"CO_EA1_41374c391138323331473846","product_type":"Thermostat","product_model":"CO_EA1","nickname":"Nash Thermostat","conn_state":1,"push_switch":1,"device_params":{"temperature":"67.3","cool_sp":71.96,"heat_sp":67.1,"working_state":"idle","temp_unit":"F","mode_sys":"auto"}},"displayName":"Nash Thermostat","UUID":"ea7070ce-4130-4bc8-ac56-6c8524d70bc4","category":1,"services":[{"displayName":"","UUID":"0000003E-0000-1000-8000-0026BB765291","constructorName":"AccessoryInformation","hiddenService":false,"primaryService":false,"characteristics":[{"displayName":"Identify","UUID":"00000014-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"Identify","value":false,"props":{"format":"bool","perms":["pw"]}},{"displayName":"Manufacturer","UUID":"00000020-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"Manufacturer","value":"Wyze","props":{"format":"string","perms":["pr"],"maxLen":64}},{"displayName":"Model","UUID":"00000021-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"Model","value":"CO_EA1","props":{"format":"string","perms":["pr"],"maxLen":64}},{"displayName":"Name","UUID":"00000023-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"Name","value":"Nash Thermostat","props":{"format":"string","perms":["pr"],"maxLen":64}},{"displayName":"Serial Number","UUID":"00000030-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"SerialNumber","value":"CO_EA1_41374c391138323331473846","props":{"format":"string","perms":["pr"],"maxLen":64}},{"displayName":"Firmware Revision","UUID":"00000052-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"FirmwareRevision","value":"0.5.25-beta.4","props":{"format":"string","perms":["pr"]}}],"optionalCharacteristics":[{"displayName":"Accessory Flags","UUID":"000000A6-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"AccessoryFlags","value":0,"props":{"format":"uint32","perms":["ev","pr"]}},{"displayName":"App Matching Identifier","UUID":"000000A4-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"AppMatchingIdentifier","value":"","props":{"format":"tlv8","perms":["pr"]}},{"displayName":"Configured Name","UUID":"000000E3-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"ConfiguredName","value":"","props":{"format":"string","perms":["ev","pr","pw"]}},{"displayName":"Hardware Finish","UUID":"0000026C-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"HardwareFinish","value":"","props":{"format":"tlv8","perms":["pr"]}},{"displayName":"Hardware Revision","UUID":"00000053-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"HardwareRevision","value":"","props":{"format":"string","perms":["pr"]}},{"displayName":"Product Data","UUID":"00000220-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"ProductData","value":"","props":{"format":"data","perms":["pr"]}},{"displayName":"Software Revision","UUID":"00000054-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"SoftwareRevision","value":"","props":{"format":"string","perms":["pr"]}}]},{"displayName":"","UUID":"0000004A-0000-1000-8000-0026BB765291","constructorName":"Thermostat","hiddenService":false,"primaryService":false,"characteristics":[{"displayName":"Current Heating Cooling State","UUID":"0000000F-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"CurrentHeatingCoolingState","value":0,"props":{"format":"uint8","perms":["ev","pr"],"minValue":0,"maxValue":2,"minStep":1,"validValues":[0,1,2]}},{"displayName":"Target Heating Cooling State","UUID":"00000033-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"TargetHeatingCoolingState","value":3,"props":{"format":"uint8","perms":["ev","pr","pw"],"minValue":0,"maxValue":3,"minStep":1,"validValues":[0,1,2,3]}},{"displayName":"Current Temperature","UUID":"00000011-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"CurrentTemperature","value":20,"props":{"format":"float","perms":["ev","pr"],"unit":"celsius","minValue":-270,"maxValue":100,"minStep":0.1}},{"displayName":"Target Temperature","UUID":"00000035-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"TargetTemperature","value":20,"props":{"format":"float","perms":["ev","pr","pw"],"unit":"celsius","minValue":10,"maxValue":38,"minStep":0.1}},{"displayName":"Temperature Display Units","UUID":"00000036-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"TemperatureDisplayUnits","value":1,"props":{"format":"uint8","perms":["ev","pr","pw"],"minValue":0,"maxValue":1,"minStep":1,"validValues":[0,1]}},{"displayName":"Cooling Threshold Temperature","UUID":"0000000D-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"CoolingThresholdTemperature","value":22.200000000000003,"props":{"format":"float","perms":["ev","pr","pw"],"unit":"celsius","minValue":10,"maxValue":35,"minStep":0.1}},{"displayName":"Heating Threshold Temperature","UUID":"00000012-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"HeatingThresholdTemperature","value":19.5,"props":{"format":"float","perms":["ev","pr","pw"],"unit":"celsius","minValue":0,"maxValue":25,"minStep":0.1}}],"optionalCharacteristics":[{"displayName":"Name","UUID":"00000023-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"Name","value":"","props":{"format":"string","perms":["pr"],"maxLen":64}},{"displayName":"Current Relative Humidity","UUID":"00000010-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"CurrentRelativeHumidity","value":0,"props":{"format":"float","perms":["ev","pr"],"unit":"percentage","minValue":0,"maxValue":100,"minStep":1}},{"displayName":"Target Relative Humidity","UUID":"00000034-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"TargetRelativeHumidity","value":0,"props":{"format":"float","perms":["ev","pr","pw"],"unit":"percentage","minValue":0,"maxValue":100,"minStep":1}},{"displayName":"Cooling Threshold Temperature","UUID":"0000000D-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"CoolingThresholdTemperature","value":10,"props":{"format":"float","perms":["ev","pr","pw"],"unit":"celsius","minValue":10,"maxValue":35,"minStep":0.1}},{"displayName":"Heating Threshold Temperature","UUID":"00000012-0000-1000-8000-0026BB765291","eventOnlyCharacteristic":false,"constructorName":"HeatingThresholdTemperature","value":0,"props":{"format":"float","perms":["ev","pr","pw"],"unit":"celsius","minValue":0,"maxValue":25,"minStep":0.1}}]}]}]
